# employees-api/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Employees API",
    "description": "Documentation automatically generated by the <b>swagger.autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "tags": [
    {
      "name": "Employees",
      "description": "Endpoints"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/employees": {
      "get": {
        "tags": [
          "Employees"
        ],
        "description": "Endpoint to get an employee.",
        "parameters": [
          {
            "name": "searchFor",
            "description": "Search type used. Ex: name, cpf, job position...",
            "type": "string",
            "in": "query"
          },
          {
            "name": "employeeData",
            "description": "Employee data.",
            "type": "string or array",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee found!",
            "schema": {
              "$ref": "#/definitions/Employee"
            }
          }
        }
      }
    },
    "/api/employees/post": {
      "post": {
        "tags": [
          "Employees"
        ],
        "description": "Endpoint to add an employee or update if it already exists.",
        "parameters": [
          {
            "name": "employee",
            "in": "body",
            "description": "Employee informations.",
            "required": true,
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AddEmployee"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Employee added successfully!"
          }
        }
      }
    },
    "/api/employees/delete": {
      "delete": {
        "tags": [
          "Employees"
        ],
        "description": "Endpoint to delete an employee.",
        "parameters": [
          {
            "name": "cpf",
            "in": "body.cpf",
            "description": "Employee cpf.",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/DeleteEmployee"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Employee successfully removed!"
          }
        }
      }
    }
  },
  "definitions": {
    "Employee": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "cpf": {
          "type": "string",
          "example": "357.029.218-56"
        },
        "position": {
          "type": "string",
          "example": "Dev Jr"
        },
        "date": {
          "type": "string",
          "example": "14/06/2019"
        },
        "uf": {
          "type": "string",
          "example": "CE"
        },
        "salary": {
          "type": "number",
          "example": 2000
        },
        "status": {
          "type": "string",
          "example": "ATIVO"
        }
      },
      "xml": {
        "name": "Employee"
      }
    },
    "AddEmployee": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Jhon Doe"
        },
        "cpf": {
          "type": "string",
          "example": "357.029.218-56"
        },
        "position": {
          "type": "string",
          "example": "Dev Jr"
        },
        "date": {
          "type": "string",
          "example": "14/06/2019"
        },
        "uf": {
          "type": "string",
          "example": "CE"
        },
        "salary": {
          "type": "number",
          "example": 2000
        },
        "status": {
          "type": "string",
          "example": "ATIVO"
        }
      },
      "required": [
        "name",
        "cpf",
        "position",
        "date",
        "uf",
        "salary",
        "status"
      ],
      "xml": {
        "name": "AddEmployee"
      }
    },
    "DeleteEmployee": {
      "type": "object",
      "properties": {
        "cpf": {
          "type": "string",
          "example": "357.029.218-56"
        }
      },
      "required": [
        "cpf"
      ],
      "xml": {
        "name": "DeleteEmployee"
      }
    }
  }
}